Refactored authentication files on server and also made a UserHomePage component, updated routes for redirect upon login to /user
