+++ client/src/app.jsx
@@ -4,6 +4,7 @@ import $ from 'jquery';
 import PodcastMain from './components/PodcastMain.jsx';
 import Login from './components/Login.jsx';
 import Signup from './components/Signup.jsx';
+import UserHomePage from './components/UserHomePage.jsx';
 import Layout from './layout/Layout.jsx';
 
 import { HashRouter as Router, Route, Switch } from 'react-router-dom';
@@ -39,6 +40,7 @@ class App extends React.Component {
             <Route name="root" exact path="/" component={() => (<PodcastMain onSearch={this.onSearch} podcasts={this.state.podcasts} /> )}/>
             <Route path="/login" component={Login} />
             <Route path="/signup" component={Signup} />
+            <Route path="/user" component={UserHomePage} />
           </Switch>
         </div>
       </Router>
