+++ server/index.js
@@ -25,7 +25,7 @@ app.use('/', express.static(path.join(__dirname + '/../client')));
 
 app.use('/', routes);
 
-const config = require('./config.js');
+const authentication = require('./auth/authentication.js');
 
 app.listen(app.port, function () {
   console.log('Example app listening on port ' + app.port);
